put data_structures/location as a separate module
