+++ internal/test_cases/geoadd_test_case.go
@@ -3,7 +3,7 @@ package test_cases
 import (
 	"strconv"
 
-	"github.com/codecrafters-io/redis-tester/internal/data_structures"
+	"github.com/codecrafters-io/redis-tester/internal/data_structures/location"
 	"github.com/codecrafters-io/redis-tester/internal/instrumented_resp_connection"
 	"github.com/codecrafters-io/redis-tester/internal/resp_assertions"
 	"github.com/codecrafters-io/tester-utils/logger"
@@ -11,7 +11,7 @@ import (
 
 type GeoAddTestCase struct {
 	Key                         string
-	Location                    data_structures.Location
+	Location                    location.Location
 	ExpectedAddedLocationsCount int
 }
