+++ internal/test_geospatial_calculate_score.go
@@ -1,7 +1,7 @@
 package internal
 
 import (
-	"github.com/codecrafters-io/redis-tester/internal/data_structures"
+	"github.com/codecrafters-io/redis-tester/internal/data_structures/location"
 	"github.com/codecrafters-io/redis-tester/internal/instrumented_resp_connection"
 	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"github.com/codecrafters-io/redis-tester/internal/resp_assertions"
@@ -26,14 +26,14 @@ func testGeospatialCalculateScore(stageHarness *test_case_harness.TestCaseHarnes
 	defer client.Close()
 
 	locationKey := testerutils_random.RandomWord()
-	locationSet := data_structures.GenerateRandomLocationSet(testerutils_random.RandomInt(2, 4))
+	locationSet := location.GenerateRandomLocationSet(testerutils_random.RandomInt(2, 4))
 	locations := locationSet.GetLocations()
 
 	// Add locations
-	for _, location := range locations {
+	for _, loc := range locations {
 		geoAddTestCase := test_cases.GeoAddTestCase{
 			Key:                         locationKey,
-			Location:                    location,
+			Location:                    loc,
 			ExpectedAddedLocationsCount: 1,
 		}
 		if err := geoAddTestCase.Run(client, logger); err != nil {
@@ -43,11 +43,11 @@ func testGeospatialCalculateScore(stageHarness *test_case_harness.TestCaseHarnes
 
 	// Check the score of each location
 	logger.Infof("Checking the scores of added locations")
-	for _, l := range locations {
+	for _, loc := range locations {
 		zscoreTestCase := test_cases.SendCommandTestCase{
 			Command:   "ZSCORE",
-			Args:      []string{locationKey, l.Name},
-			Assertion: resp_assertions.NewFloatingPointBulkStringAssertion(float64(l.GetGeoCode()), 0),
+			Args:      []string{locationKey, loc.Name},
+			Assertion: resp_assertions.NewFloatingPointBulkStringAssertion(float64(loc.GetGeoCode()), 0),
 		}
 		if err := zscoreTestCase.Run(client, logger); err != nil {
 			return err
