+++ internal/test_geospatial_decode_coordinates.go
@@ -2,6 +2,7 @@ package internal
 
 import (
 	"github.com/codecrafters-io/redis-tester/internal/data_structures"
+	"github.com/codecrafters-io/redis-tester/internal/data_structures/location"
 	"github.com/codecrafters-io/redis-tester/internal/instrumented_resp_connection"
 	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"github.com/codecrafters-io/redis-tester/internal/test_cases"
@@ -26,16 +27,16 @@ func testGeospatialDecodeCoordinates(stageHarness *test_case_harness.TestCaseHar
 
 	locationKey := testerutils_random.RandomWord()
 
-	locationSet := data_structures.GenerateRandomLocationSet(testerutils_random.RandomInt(2, 4))
+	locationSet := location.GenerateRandomLocationSet(testerutils_random.RandomInt(2, 4))
 
 	// We use ZADD to add the locations so the user implementation requires that they decode
 	// co-ordinates from the score
-	for _, location := range locationSet.GetLocations() {
+	for _, loc := range locationSet.GetLocations() {
 		zaddTestCase := test_cases.ZaddTestCase{
 			Key: locationKey,
 			Member: data_structures.SortedSetMember{
-				Name:  location.Name,
-				Score: float64(location.GetGeoCode()),
+				Name:  loc.Name,
+				Score: float64(loc.GetGeoCode()),
 			},
 			ExpectedAddedMembersCount: 1,
 		}
@@ -44,7 +45,7 @@ func testGeospatialDecodeCoordinates(stageHarness *test_case_harness.TestCaseHar
 		}
 	}
 
-	geoPosTestCase := test_cases.NewGeoPosTestCase(locationKey, true)
+	geoPosTestCase := test_cases.NewGeoPosTestCase(locationKey, false)
 	geoPosTestCase.AddLocations(locationSet.GetLocations())
 
 	return geoPosTestCase.Run(client, logger)
